feat(commute): 디렉토리 구조 변경 및 예외규칙 추가/수정/삭제 기능구현
