+++ src/main/java/com/finalproj/orbitflow/approval/attendanceRecord/entity/AttendanceRecord.java
@@ -11,7 +11,7 @@ package com.finalproj.orbitflow.approval.attendanceRecord.entity;
 
 import com.finalproj.orbitflow.approval.document.entity.Document;
 import com.finalproj.orbitflow.approval.document.enums.DocumentStatus;
-import com.finalproj.orbitflow.attendance.leave.entity.LeaveType;
+import com.finalproj.orbitflow.leave.leaveType.entity.LeaveType;
 import com.finalproj.orbitflow.global.common.BaseEntity;
 import com.finalproj.orbitflow.hr.company.entity.Company;
 import com.finalproj.orbitflow.hr.employee.entity.Employee;
