+++ src/main/java/com/finalproj/orbitflow/attendance/attendance/repository/AttendanceRepository.java
@@ -1,0 +1,19 @@
+package com.finalproj.orbitflow.attendance.attendance.repository;
+
+import com.finalproj.orbitflow.attendance.attendance.entity.Attendance;
+import org.springframework.data.jpa.repository.JpaRepository;
+import java.time.LocalDate;
+import java.util.Optional;
+
+public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
+
+    // 회사 ID와 사원 ID, 날짜로 조회
+    Optional<Attendance> findByCompanyIdAndEmployeeIdAndWorkDate(Long companyId, Long employeeId, LocalDate workDate);
+
+    // 오늘 기록 조회 편의 메서드
+    default Optional<Attendance> findToday(Long companyId, Long employeeId) {
+        return findByCompanyIdAndEmployeeIdAndWorkDate(companyId, employeeId, LocalDate.now());
+    }
+
+    Optional<Attendance> findByEmployeeIdAndWorkDate(Long employeeId, LocalDate now);
+}
