+++ src/main/java/com/finalproj/orbitflow/attendance/attendanceRule/dto/AttRuleResDto.java
@@ -1,0 +1,26 @@
+package com.finalproj.orbitflow.attendance.attendanceRule.dto;
+
+import com.finalproj.orbitflow.attendance.attendanceRule.entity.AttendanceRule;
+
+import java.time.LocalTime;
+
+/**
+ * Please explain the class!!!
+ *
+ * @author : rlagkdus
+ * @filename : AttendanceRuleResDto
+ * @since : 2025. 12. 18. 목요일
+ */
+public record AttRuleResDto(
+        LocalTime defaultStartTime,
+        LocalTime defaultEndTime,
+        Integer defaultBreakMinutes
+) {
+    public AttRuleResDto(AttendanceRule rule) {
+        this(
+                rule.getDefaultStartTime(),
+                rule.getDefaultEndTime(),
+                rule.getDefaultBreakMinutes()
+        );
+    }
+}
